propagade to custom packages  app - all business logic  testframework - all with running tests  tests - test copy of business logic for unit tests
