+++ HW06-Annotation/src/main/java/com/akproductions/Hw06_main.java
@@ -1,9 +1,11 @@
-package com.testframework;
+package com.akproductions;
+
+import com.akproductions.testframework.TestFramework;
 
 public class Hw06_main {
 
     public static void main(String[] args) throws Exception {
-        String ClassToTest = "com.akproductions.MessageBuilderImpl";
+        String ClassToTest = "com.akproductions.tests.MessageBuilderImplTest";
         System.out.println((new TestFramework(ClassToTest)).doTest());
     }
 }
