--- HW06-Annotation/src/main/java/com/akproductions/MessageTemplateProvider.java
@@ -1,5 +1,0 @@
-package com.akproductions;
-
-public class MessageTemplateProvider {
-  String getMessageTemplate(String templateName){return templateName;};
-}
