+++ HW06-Annotation/src/main/java/com/akproductions/app/MessageBuilder.java
@@ -1,4 +1,4 @@
-package com.akproductions;
+package com.akproductions.app;
 
 public interface MessageBuilder {
   String buildMessage(String templateName, String text, String signature);
