+++ HW06-Annotation/src/main/java/com/akproductions/app/MessageTemplateProvider.java
@@ -1,0 +1,5 @@
+package com.akproductions.app;
+
+public class MessageTemplateProvider {
+  public String getMessageTemplate(String templateName){return templateName;};
+}
