+++ HW06-Annotation/src/main/java/com/akproductions/app/TemplateNotFoundException.java
@@ -1,4 +1,4 @@
-package com.akproductions;
+package com.akproductions.app;
 
 public class TemplateNotFoundException extends RuntimeException {
 }
