+++ HW06-Annotation/src/main/java/com/akproductions/testframework/annotations/Before.java
@@ -1,4 +1,4 @@
-package com.akproductions;
+package com.akproductions.testframework.annotations;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
