+++ HW06-Annotation/src/main/java/com/akproductions/testframework/annotations/DisplayName.java
@@ -1,4 +1,4 @@
-package com.akproductions;
+package com.akproductions.testframework.annotations;
 
 public @interface DisplayName {
     String name() default "";
