+++ HW06-Annotation/src/main/java/com/akproductions/tests/MessageBuilderImplTest.java
@@ -1,0 +1,98 @@
+package com.akproductions.tests;
+
+import com.akproductions.app.MessageBuilder;
+import com.akproductions.app.MessageTemplateProvider;
+import com.akproductions.app.TemplateNotFoundException;
+import com.akproductions.testframework.annotations.After;
+import com.akproductions.testframework.annotations.Before;
+import com.akproductions.testframework.annotations.DisplayName;
+import com.akproductions.testframework.annotations.Test;
+
+public class MessageBuilderImplTest implements MessageBuilder {
+
+    //переменные для тестов reflection
+    int var1;
+    int var2;
+    int getMessageTemplateCalled;
+
+    //метод для тестов reflection
+    String getValue(){
+        return "var1="+ var1+ ", var2="+ var2;
+    }
+
+    public static final String TEMPLATE_NAME = "AnyTemplate";
+    public static final String TEMPLATE_TEXT = "Hi!\n %s \n With best regards, %s";
+    public static final String MESSAGE_TEXT = "How you doing?";
+    public static final String SIGN = "Vasya";
+    private MessageBuilder messageBuilder;
+    private final MessageTemplateProvider templateProvider;
+
+    public MessageBuilderImplTest() {
+        templateProvider=new MessageTemplateProvider();
+        getMessageTemplateCalled=0;
+    }
+
+    public MessageBuilderImplTest(MessageTemplateProvider templateProvider) {
+        this.templateProvider = templateProvider;
+    }
+
+    @Override
+    public String buildMessage(String templateName, String text, String signature) {
+        String messageTemplate = templateProvider.getMessageTemplate(templateName);
+        if (messageTemplate == null || messageTemplate.isEmpty()) {
+            throw new TemplateNotFoundException();
+        }
+        getMessageTemplateCalled++;
+        return String.format(messageTemplate, text, signature);
+    }
+
+    //метод для тестов reflection
+    @Before
+    private void doBefore()
+    {
+        System.out.println("Before called!!! var1=" + var1+ ", var2="+ var2);
+    }
+    //метод для тестов reflection
+    @After
+    private void doAfter()
+    {
+        System.out.println("After called!!! var1=" + var1+ ", var2="+ var2);
+    }
+
+    //метод для тестов reflection
+    @Test
+    private void doTest()
+    {
+        var1 += var2;
+        System.out.println("Test called!!! var1=" + var1+ ", var2="+ var2);
+    }
+
+    @Test
+    @DisplayName(name = "buildMessage1 - проверить что корректно работает логика метода buildMessage")
+    void buildMessage1() throws Exception {
+        //when(provider.getMessageTemplate(TEMPLATE_NAME)).thenReturn(TEMPLATE_TEXT);
+        String actualMessage = buildMessage(TEMPLATE_TEXT,MESSAGE_TEXT, SIGN);
+        if (!(String.format(TEMPLATE_TEXT, MESSAGE_TEXT, SIGN).contentEquals(actualMessage)))
+        {
+            throw new Exception("[FATAL ERROR] buildMessage working incorrect!!!");
+        }
+    }
+
+    @Test
+    @DisplayName(name = " buildMessage2 - проверить что хотя бы раз вызывается метод getMessageTemplate")
+    void buildMessage2() throws Exception {
+        String actualMessage = buildMessage(TEMPLATE_TEXT, MESSAGE_TEXT, SIGN);
+        if (this.getMessageTemplateCalled == 0)
+        {
+            throw new Exception("[FATAL ERROR] getMessageTemplate not invoked during this test!!!");
+        }
+    }
+
+    @Test
+    @DisplayName(name = " buildMessage3 - проверить генерацию исключения TemplateNotFoundException")
+    void buildMessage3() throws Exception {
+        //when(provider.getMessageTemplate(TEMPLATE_NAME)).thenReturn(TEMPLATE_TEXT);
+        throw new Exception("[ERROR] " + TemplateNotFoundException.class.toString());
+    }
+
+}
